+++ R/bold_identify.R
@@ -63,37 +63,39 @@
 #' }
 #'
 #' @export
-bold_identify <- function(sequences, db = 'COX1', response = FALSE, keepSeq = TRUE, ...) {
+bold_identify <- function(
+    sequences,
+    db = c("COX1", "COX1_SPECIES", "COX1_SPECIES_PUBLIC", "COX1_L640bp"),
+    response = FALSE, keepSeq = TRUE, ...){
+  if (missing(sequences)) stop("argument 'sequences' is missing, with no default")
   # make sure sequences are character before the queries
   if (is.list(sequences) && all(lengths(sequences) == 1)) {
     sequences <- c(sequences, recursive = TRUE)
   }
-  assert(sequences, "character")
+  b_assert(sequences, "character", check.length = 0L)
   if (!missing(db)) {
-    assert(db, "character", check.length.is1 = TRUE)
-    if (!db %in% c("COX1", "COX1_SPECIES", "COX1_SPECIES_PUBLIC", "COX1_L640bp")) {
-      stop("'db' must be one of 'COX1', 'COX1_SPECIES', 'COX1_SPECIES_PUBLIC' or 'COX1_L640bp'")
-    }
+    db <- b_assert_db(db)
+  } else {
+    db <- "COX1"
   }
-  assert(response, "logical")
-  assert(keepSeq, "logical")
+  response <- b_assert_logical(response)
+  keepSeq <- b_assert_logical(keepSeq)
   # loop over sequences since the API only accepts one at a time
   lapply(sequences, function(seq){
     if (nchar(seq) < 80) {
-      res <- list(response = NULL, warning = "Sequence  must be at least 80 bp.")
+      # API return NA but on website you would get that error
+      res <- list(response = NULL, warning = "Sequence must be at least 80 bp")
     } else if (!grepl("^[A-z|*|-|+|.]*$", seq)) {
-      res <- list(response = NULL, warning = "Sequence contains invalid characters.")
+      # API return NA but on website you would get that error
+      res <- list(response = NULL, warning = "Sequence contains invalid characters")
     } else {
-      res <- get_response(
-        args = bc(list(sequence = seq, db = db)),
-        url = b_url('Ids_xml'),
-        contentType = 'text/xml'
-      )
+      res <- b_GET(query = b_rm_empty(list(sequence = seq, db = db)),
+                   api = 'Ids_xml', check = TRUE, contentType = 'text/xml')
     }
     if (response) {
       res
     } else {
-      out <- if (length(res$response)) .parse_identify_xml(res) else NA
+      out <- if (length(res$response)) b_parse_identify(res) else NA
       # add input sequence as attribute to prevent mix up if 'sequences' wasn't named
       if (keepSeq) {
         attr(out, "sequence") <- seq
@@ -104,7 +106,7 @@ bold_identify <- function(sequences, db = 'COX1', response = FALSE, keepSeq = TR
   })
 }
 
-.parse_identify_xml <- function(res){
+b_parse_identify <- function(res){
   cNames <- c('ID','sequencedescription','database','citation',
               'taxonomicidentification','similarity','specimen_url',
               'specimen_country', 'specimen_lat','specimen_lon')