+++ R/bold_identify_taxonomy.R
@@ -38,13 +38,13 @@ methods::setMethod(
   f = "bold_identify_taxonomy",
   signature = signature("list"),
   definition = function(x, taxOnly = TRUE) {
-    if (!length(x)) stop("argument `x` is empty")
-    assert(taxOnly, "logical")
+    b_assert_length(x, len = 0L, name = "x")
+    taxOnly <- b_assert_logical(taxOnly, name = "taxOnly")
     hasID <- vapply(x, \(x) "ID" %in% colnames(x), NA)
     if (all(!hasID)) {
       stop("no column 'ID' found in x")
     } else {
-      lapply(x, .bold_identify_taxonomy, taxOnly = taxOnly)
+      lapply(x, b_identify_taxonomy, taxOnly = taxOnly)
     }
   }
 )
@@ -53,9 +53,8 @@ methods::setMethod(
   f = "bold_identify_taxonomy",
   signature = signature("matrix"),
   definition = function(x, taxOnly = TRUE) {
-    if (!length(x)) stop("argument `x` is empty")
-    assert(taxOnly, "logical")
-    .bold_identify_taxonomy(x, taxOnly = taxOnly)
+    b_assert_length(x, len = 0L, name = "x")
+    b_identify_taxonomy(x, taxOnly = b_assert_logical(taxOnly, name = "taxOnly"))
   }
 )
 #' @rdname bold_identify_taxonomy
@@ -63,9 +62,8 @@ methods::setMethod(
   f = "bold_identify_taxonomy",
   signature = signature("data.frame"),
   definition = function(x, taxOnly = TRUE) {
-    if (!length(x)) stop("argument `x` is empty")
-    assert(taxOnly, "logical")
-    .bold_identify_taxonomy(x, taxOnly = taxOnly)
+    b_assert_length(x, len = 0L, name = "x")
+    b_identify_taxonomy(x, taxOnly = b_assert_logical(taxOnly, name = "taxOnly"))
   }
 )
 #' @rdname bold_identify_taxonomy
@@ -76,7 +74,7 @@ methods::setMethod(
       stop("argument 'x' is missing, with no default")
   }
 )
-.bold_identify_taxonomy <- function(x, taxOnly){
+b_identify_taxonomy <- function(x, taxOnly){
   if (!any(colnames(x) == "ID")) {
     warning("No column 'ID' found, skipped", call. = FALSE, immediate. = TRUE)
     data.frame(x)
@@ -87,7 +85,7 @@ methods::setMethod(
     tax <- bold_specimens(ids = unique(IDs))
     if (taxOnly) {
       # only returns the taxonomic data
-      tax <- tax[, c("processid", grep("_taxID|_name", names(tax), value = TRUE))]
+      tax <- tax[, c("processid", names(tax)[b_detect(names(tax), "_taxID|_name")])]
     }
     tax[tax == ""] <- NA
     # remove empty columns