+++ R/bold_tax_id.R
@@ -53,7 +53,7 @@ bold_tax_id <- function(id, dataTypes = "basic", includeTree = FALSE,
                         response = FALSE, ...) {
   .Deprecated("bold_tax_id2")
   #-- arguments check
-  # no need to do the call if id is NA
+  if (missing(id)) stop("argument 'id' is missing, with no default")
   if (!grepl("true|false", includeTree, ignore.case = TRUE))
     warning("'includeTree' should be either TRUE or FALSE.")
   if (!inherits(id, c("character", "numeric", "integer")))
@@ -69,6 +69,7 @@ bold_tax_id <- function(id, dataTypes = "basic", includeTree = FALSE,
     URL <- b_url("API_Tax/TaxonData")
     #-- fetch data from the api
     res <- lapply(`names<-`(id, id), function(x) {
+      # no need to do the call if id is NA
       if (is.na(x))
         data.frame(input = NA, noresults = NA)
       else
@@ -77,7 +78,7 @@ bold_tax_id <- function(id, dataTypes = "basic", includeTree = FALSE,
     if (response) {
       out <- res
     } else {
-      out <- setrbind(mapply(FUN = .process_response_dep, x = res, y = id,
+      out <- b_rbind(mapply(FUN = .process_response_dep, x = res, y = id,
                              MoreArgs = list(z = includeTree, w = dataTypes),
                              SIMPLIFY = FALSE))
       if (NCOL(out) == 1) {
@@ -105,7 +106,7 @@ bold_tax_id <- function(id, dataTypes = "basic", includeTree = FALSE,
     if (!is.null(names(out))) {
       df <- data.frame(out, stringsAsFactors = FALSE)
     } else {
-      df <- setrbind(lapply(out, data.frame, stringsAsFactors = FALSE))
+      df <- b_rbind(lapply(out, data.frame, stringsAsFactors = FALSE))
     }
     row.names(df) <- NULL
     if ("parentid" %in% names(df)) df <- df[order(df[,"parentid"]),]
@@ -121,33 +122,20 @@ bold_tax_id <- function(id, dataTypes = "basic", includeTree = FALSE,
   return(out)
 }
 .check_dataTypes_dep <- function(x){
-  x <- stringi::stri_split_fixed(x, ",", simplify = TRUE)
+  x <- b_split(x, ",", fixed = TRUE, simplify = TRUE)
   # corrects for the json typo in case the option is taken from a previous query
-  if (any(x == "depositories")) {
-    x[x == "depositories"] <- "depository"
-  }
+  x[x == "depositories"] <- "depository"
   if (length(x) > 1 && any(x == "all")) {
     x <- "all"
   } else {
-    wrongType <- !x %in% c(
-      "basic",
-      "stats",
-      "geo",
-      "images",
-      "sequencinglabs",
-      "depository",
-      "thirdparty",
-      "all"
-    )
+    wrongType <- !x %in% b_dataTypes
     if (any(wrongType)) {
-      wt <- toStr(x[wrongType], quote = TRUE)
-      warning(wt,
+      warning(b_ennum(x[wrongType], quote = TRUE),
               if (sum(wrongType) > 1) " are not valid data types"
               else " is not a valid data type",
               if (!all(wrongType)) " and will be skipped." else ".",
-              "\nThe possible data types are:",
-              "\n  - basic\n  - stats\n  - geo\n  - images",
-              "\n  - sequencinglabs\n  - depository\n  - thirdparty\n  - all")
+              "\nChoices are:",
+              b_ennum(b_dataTypes, "or", TRUE))
       x <- x[!wrongType]
     }
     x <- paste(x, collapse = ",")