+++ tests/testthat/test-bold_tax_name.R
@@ -9,7 +9,6 @@ test_that("bold_tax_name returns the correct object", {
   expect_is(test$input, "character")
   expect_is(test$taxid, "integer")
 })
-
 test_that("bold_tax_name returns the correct object (multiple names)", {
   skip_on_cran()
   vcr::use_cassette("bold_tax_name", {
@@ -50,7 +49,6 @@ test_that("bold_tax_name returns the correct object (when taxon not found)", {
   expect_is(test$taxid, "integer")
   expect_true(is.na(test$taxid))
 })
-
 test_that("bold_tax_name returns the correct object (response)", {
   skip_on_cran()
   vcr::use_cassette("bold_tax_name", {
@@ -66,7 +64,6 @@ test_that("bold_tax_name returns the correct object (response)", {
   expect_is(test$warning, "character")
   expect_equal(test$warning, "")
 })
-
 test_that("bold_tax_name 'fuzzy' param works", {
   skip_on_cran()
   vcr::use_cassette("bold_tax_name", {
@@ -78,7 +75,7 @@ test_that("bold_tax_name 'fuzzy' param works", {
   expect_gt(NROW(test), NROW(test_not))
 })
 
-test_that("bold_tax_name works when 'name' containts single quotes or ends with a dot or a closing parenthesis", {
+test_that("bold_tax_name works when  'name' containts single quotes", {
   skip_on_cran()
   # see issue 84
   vcr::use_cassette("bold_tax_name", {
@@ -94,22 +91,25 @@ test_that("bold_tax_name works when 'name' containts single quotes or ends with 
   expect_equal(test$input, test$taxon)
   expect_true(all(test$input == test$taxonrep, na.rm = TRUE))
 })
-
 test_that("bold_tax_name fails well", {
   # name required
-  expect_error(bold_tax_name(), "argument \"name\" is missing, with no default")
+  expect_error(bold_tax_name(), "argument 'name' is missing, with no default")
   # catch wrong type param inputs
   expect_error(bold_tax_name(name = "Diplur", response = 5),
-               "'response' must be of class logical.")
+               "'response' should be one of TRUE or FALSE")
   expect_error(bold_tax_name(name = "Diplur", fuzzy = 5),
-               "'fuzzy' must be of class logical.")
-  expect_error(bold_tax_name(name = "Diplur", tax_division = 5),
-               "'tax_division' must be of class character.")
-  expect_error(bold_tax_name(name = "Diplur", tax_division = "Mushrooms"),
-               "'tax_division' must be one or more of Animalia, Protista, Fungi and Plantae")
+               "'fuzzy' should be one of TRUE or FALSE")
+  expect_error(
+    bold_tax_name(name = "Diplur", tax_division = 5),
+    "'tax_division' must be of class character"
+  )
+  expect_error(
+    bold_tax_name(name = "Diplur", tax_division = "Mushrooms"),
+    "is not a valid tax_division"
+  )
   expect_error(bold_tax_name(name = "Diplur", tax_rank = 5),
-               "'tax_rank' must be of class character.")
+               "'tax_rank' must be of class character")
   expect_error(bold_tax_name(name = "Diplur",
                              tax_rank = c("genus", "notArank")),
-               "Invalid tax_rank name.")
+               "is not a valid tax_rank")
 })