+++ tests/testthat/test-bold_trace.R
@@ -1,3 +1,5 @@
+context("bold_trace")
+# vcr::vcr_configure(write_disk_path = "D:\\02_rOpenSci\\bold\\tests\\bold_trace_files")
 test_that("bold_trace returns the correct object", {
   skip_on_cran()
   dest.dir <- vcr::vcr_configuration()$write_disk_path
@@ -11,7 +13,6 @@ test_that("bold_trace returns the correct object", {
   expect_length(test$ab1, 6L)
   expect_is(test$args, "character")
 })
-
 test_that("bold_trace returns the correct object (with ids)", {
   skip_on_cran()
   dest.dir <- vcr::vcr_configuration()$write_disk_path
@@ -25,7 +26,6 @@ test_that("bold_trace returns the correct object (with ids)", {
   expect_length(test$ab1, 3L)
   expect_is(test$args, "character")
 })
-
 test_that("print.bold_trace prints properly", {
   skip_on_cran()
   dest.dir <- vcr::vcr_configuration()$write_disk_path
@@ -38,7 +38,6 @@ test_that("print.bold_trace prints properly", {
   expect_true(sum(!nzchar(test)) == 2)
   expect_true(sum(grepl(".ab1$", test)) == 6)
 })
-
 test_that("bold_read_trace works properly", {
   skip_on_cran()
   dest.dir <- vcr::vcr_configuration()$write_disk_path
@@ -64,11 +63,10 @@ test_that("bold_read_trace works properly", {
   expect_length(test_trace, 1)
   expect_length(test_trace[[1]], 0)
 })
-
 test_that("bold_trace fails well", {
   expect_error(bold_trace(), "You must provide a non-empty value to at least one of")
   expect_error(bold_trace(taxon = ''), "You must provide a non-empty value to at least one of")
-  expect_error(bold_trace(taxon = 5, geo = 1), "'taxon' and 'geo' must be of class character.")
-  expect_error(bold_trace(taxon = 'Coelioxys', overwrite = "true"), "'overwrite' must be of class logical.")
-  expect_error(bold_trace(taxon = 'Coelioxys', dest = TRUE), "'dest' must be of class character.")
+  expect_error(bold_trace(taxon = 5, geo = 1), "'taxon' and 'geo' must be of class character")
+  expect_error(bold_trace(taxon = 'Coelioxys', overwrite = 5), "'overwrite' should be one of TRUE or FALSE")
+  expect_error(bold_trace(taxon = 'Coelioxys', dest = TRUE), "'dest' must be of class character")
 })